api/requirements.txt
fastapi[standard]
pydantic
radis
rabbitmq
pika
python-dotenv